Use :pgroup option when starting subprocesses
This makes it easier and faster to kill spawned git processes.
Repetitive `pgrep` command for collecting PIDs is no longer required.
Unfortunately, :pgroup option is not supported on Ruby 1.8 so we can't
remove the old implementation of `killall`.

However, this turned out to have negative effect on responsiveness of
Vim on Linux when user tries to interrupt the running installation.

Need further investigation.